refac

+++ pacman-ui-fx/src/main/java/de/amr/games/pacman/ui/fx/_3d/scene/PlayScene3D.java
@@ -121,7 +121,7 @@ public class PlayScene3D implements GameScene {
 	}
 
 	@Override
-	public void resizeToHeight(float height) {
+	public void resizeToHeight(double height) {
 		// nothing to do
 	}
